+++ src/transporters/amqp.js
@@ -6,6 +6,7 @@
 
 "use strict";
 
+const url = require("url");
 const Promise		= require("bluebird");
 const Transporter 	= require("./base");
 const { isPromise }	= require("../utils");
@@ -73,6 +74,19 @@ class AmqpTransporter extends Transporter {
 		if (typeof opts.consumeOptions !== "object")
 			opts.consumeOptions = {};
 
+		opts.autoDeleteQueues =
+			opts.autoDeleteQueues === true ? 2*60*1000 :
+				typeof opts.autoDeleteQueues === "number" ? opts.autoDeleteQueues :
+					opts.autoDeleteQueues === false ? -1 :
+						-1; // Eventually we could change default
+
+		// Support for multiple URLs (clusters)
+		opts.url = Array.isArray(opts.url)
+			? opts.url
+			: !opts.url
+				? ['']
+				: opts.url.split(';').filter(s => !!s)
+
 		super(opts);
 
 		this.hasBuiltInBalancer = true;
@@ -89,8 +103,18 @@ class AmqpTransporter extends Transporter {
 	 *
 	 * @memberof AmqpTransporter
 	 */
-	connect() {
-		return new Promise((resolve, reject) => {
+	connect(errorCallback) {
+		return new Promise((_resolve, _reject) => {
+			let _isResolved = false;
+			const resolve = () => {
+				_isResolved = true;
+				_resolve();
+			};
+			const reject = (err) => {
+				_reject(err);
+				if (_isResolved) errorCallback(err);
+			};
+
 			let amqp;
 			try {
 				amqp = require("amqplib");
@@ -99,7 +123,16 @@ class AmqpTransporter extends Transporter {
 				this.broker.fatal("The 'amqplib' package is missing. Please install it with 'npm install amqplib --save' command.", err, true);
 			}
 
-			amqp.connect(this.opts.url, this.opts.socketOptions)
+			// Pick url
+			this.connectAttempt = (this.connectAttempt||0)+1;
+			const urlIndex = (this.connectAttempt-1) % this.opts.url.length;
+			const uri = this.opts.url[urlIndex];
+			const urlParsed = url.parse(uri);
+
+			amqp.connect(uri, Object.assign({},
+				(this.opts.socketOptions || {}),
+				{servername: urlParsed.hostname}
+			))
 				.then(connection => {
 					this.connection = connection;
 					this.logger.info("AMQP is connected.");
@@ -107,17 +140,20 @@ class AmqpTransporter extends Transporter {
 					/* istanbul ignore next*/
 					connection
 						.on("error", (err) => {
-							this.connected = false;
-							reject(err);
-							this.logger.error("AMQP connection error.");
+							// No need to reject here since close event will be fired after
+							// if not connected at all connection promise will be rejected
+							// this.connected = false;
+							// reject(err);
+							this.logger.error("AMQP connection error.", err);
 						})
 						.on("close", (err) => {
 							this.connected = false;
-							reject(err);
-							if (!this.connectionDisconnecting)
+							if (!this.connectionDisconnecting) {
 								this.logger.error("AMQP connection is closed.");
-							else
+								reject(err);
+							} else {
 								this.logger.info("AMQP connection is closed gracefully.");
+							}
 						})
 						.on("blocked", (reason) => {
 							this.logger.warn("AMQP connection is blocked.", reason);
@@ -130,7 +166,6 @@ class AmqpTransporter extends Transporter {
 						.createChannel()
 						.then((channel) => {
 							this.channel = channel;
-							this.onConnected().then(resolve);
 							this.logger.info("AMQP channel is created.");
 
 							channel.prefetch(this.opts.prefetch);
@@ -138,17 +173,19 @@ class AmqpTransporter extends Transporter {
 							/* istanbul ignore next*/
 							channel
 								.on("close", () => {
-									this.connected = false;
 									this.channel = null;
-									reject();
+									// No need to reject here since close event on connection will handle
+									// this.connected = false;
+									// reject();
 									if (!this.channelDisconnecting)
 										this.logger.warn("AMQP channel is closed.");
 									else
 										this.logger.info("AMQP channel is closed gracefully.");
 								})
 								.on("error", (err) => {
-									this.connected = false;
-									reject(err);
+									// No need to reject here since close event will be fired after
+									// this.connected = false;
+									// reject(err);
 									this.logger.error("AMQP channel error.", err);
 								})
 								.on("drain", () => {
@@ -157,7 +194,10 @@ class AmqpTransporter extends Transporter {
 								.on("return", (msg) => {
 									this.logger.warn("AMQP channel returned a message.", msg);
 								});
+
+							return this.onConnected();
 						})
+						.then(resolve)
 						.catch((err) => {
 							/* istanbul ignore next*/
 							this.logger.error("AMQP failed to create channel.");
@@ -209,20 +249,28 @@ class AmqpTransporter extends Transporter {
 	 *
 	 * @memberof AmqpTransporter
 	 */
-	_getQueueOptions(packetType) {
+	_getQueueOptions(packetType, balancedQueue) {
 		let packetOptions;
 		switch(packetType) {
 			// Requests and responses don't expire.
 			case PACKET_REQUEST:
+				packetOptions = this.opts.autoDeleteQueues >= 0 && !balancedQueue
+					? { expires: this.opts.autoDeleteQueues }
+					: {};
+				break;
 			case PACKET_RESPONSE:
-				packetOptions = {};
+				packetOptions = this.opts.autoDeleteQueues >= 0
+					? { expires: this.opts.autoDeleteQueues }
+					: {};
 				break;
 
 			// Consumers can decide how long events live
 			// Load-balanced/grouped events
 			case PACKET_EVENT + "LB":
 			case PACKET_EVENT:
-				packetOptions = {};
+				packetOptions = this.opts.autoDeleteQueues >= 0
+					? { expires: this.opts.autoDeleteQueues }
+					: {};
 				// If eventTimeToLive is specified, add to options.
 				if (this.opts.eventTimeToLive)
 					packetOptions.messageTtl = this.opts.eventTimeToLive;
@@ -362,7 +410,7 @@ class AmqpTransporter extends Transporter {
 	 */
 	subscribeBalancedRequest(action) {
 		const queue = `${this.prefix}.${PACKET_REQUEST}B.${action}`;
-		return this.channel.assertQueue(queue, this._getQueueOptions(PACKET_REQUEST))
+		return this.channel.assertQueue(queue, this._getQueueOptions(PACKET_REQUEST, true))
 			.then(() => this.channel.consume(
 				queue,
 				this._consumeCB(PACKET_REQUEST, true),
@@ -379,7 +427,7 @@ class AmqpTransporter extends Transporter {
 	 */
 	subscribeBalancedEvent(event, group) {
 		const queue = `${this.prefix}.${PACKET_EVENT}B.${group}.${event}`;
-		return this.channel.assertQueue(queue, this._getQueueOptions(PACKET_EVENT + "LB"))
+		return this.channel.assertQueue(queue, this._getQueueOptions(PACKET_EVENT + "LB", true))
 			.then(() => this.channel.consume(
 				queue,
 				this._consumeCB(PACKET_EVENT, true),