+++ test/services/greeter.es6.service.js
@@ -1,7 +1,3 @@
-/**
- * IT IS NOT WORKING WITH THE LATEST JEST VERSION
- *
- */
 const Service = require("../../src/service");
 
 class GreeterService extends Service {